qba-174: stabilization

+++ PXQBA/Business/Macmillan.PXQBA.Business.Services/QuestionManagementService.cs
@@ -180,9 +180,10 @@ namespace Macmillan.PXQBA.Business.Services
             Question question = GetQuestion(course, questionId);
    
 
-            if (!string.IsNullOrEmpty(question.DraftFrom) ||
+            if (question != null && 
+                ((!string.IsNullOrEmpty(question.DraftFrom) && !question.IsPublishedFromDraft) ||
                 (question.Version == 1 && string.IsNullOrEmpty(question.DuplicateFrom) && string.IsNullOrEmpty(question.DuplicateFromShared))
-                )
+                ))
             {
                 questionCommands.DeleteQuestion(course.QuestionRepositoryCourseId, questionId);
                 return true;
